Summary
started to get background colors changed(also made it easy to change if we want later) and getting all input boxes so they did not have clear backgrounds... save location switched to a combo box along with other little changes. need to go over it and change backgrounds of all the text fields due to, for whatever reason, having a slight different color.